Tratamento do deleteById

+++ src/main/java/br/com/gunthercloud/coursejava/services/UserService.java
@@ -2,8 +2,10 @@ package br.com.gunthercloud.coursejava.services;
 
 import br.com.gunthercloud.coursejava.entities.User;
 import br.com.gunthercloud.coursejava.repositories.UserRepository;
+import br.com.gunthercloud.coursejava.services.exceptions.DatabaseException;
 import br.com.gunthercloud.coursejava.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -27,7 +29,14 @@ public class UserService {
         return repository.save(obj);
     }
     public void delete(Long id) {
-        repository.deleteById(id);
+        //repository.findById(id).orElseThrow(() -> new ResourceNotFoundException(id));
+        if(repository.findById(id).isEmpty()) throw new ResourceNotFoundException(id);
+        try {
+            repository.deleteById(id);
+        }
+        catch(DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
